# specs/001-image-gen-agent/contracts/api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Image Generation Agent API",
    "description": "Discord ベースの画像生成エージェント用 API。プロンプト生成、設定選定、画像生成リクエストの管理、履歴照会などを提供します。",
    "version": "1.0.0",
    "contact": {
      "name": "Diffuse Pilot Project",
      "url": "https://github.com/hiyori-akane/diffuse-pilot"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "ローカル開発サーバー"
    },
    {
      "url": "https://api.diffuse-pilot.example.com",
      "description": "本番環境（例）"
    }
  ],
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "summary": "グローバル設定を取得",
        "description": "指定されたサーバー（guild）とユーザーのグローバル設定を返します。",
        "operationId": "getGlobalSettings",
        "tags": ["Settings"],
        "parameters": [
          {
            "name": "guild_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Discord サーバー（guild）ID"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Discord ユーザー ID（省略時はサーバーデフォルト）"
          }
        ],
        "responses": {
          "200": {
            "description": "グローバル設定",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSettingsResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "グローバル設定を更新",
        "description": "サーバー（guild）またはユーザーのグローバル設定を更新します。",
        "operationId": "updateGlobalSettings",
        "tags": ["Settings"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalSettingsInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "設定が更新されました",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSettingsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GlobalSettingsInput": {
        "type": "object",
        "required": ["guild_id"],
        "properties": {
          "guild_id": {
            "type": "string"
          },
          "user_id": {
            "type": "string",
            "nullable": true
          },
          "default_model": {
            "type": "string",
            "nullable": true
          },
          "default_lora_list": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "weight": {
                  "type": "number",
                  "format": "float"
                }
              }
            },
            "nullable": true
          },
          "default_prompt_suffix": {
            "type": "string",
            "nullable": true
          },
          "default_sd_params": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          }
        }
      },
      "GlobalSettingsResponse": {
        "type": "object",
        "properties": {
          "settings_id": {
            "type": "string",
            "format": "uuid"
          },
          "guild_id": {
            "type": "string"
          },
          "user_id": {
            "type": "string",
            "nullable": true
          },
          "default_model": {
            "type": "string",
            "nullable": true
          },
          "default_lora_list": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "weight": {
                  "type": "number",
                  "format": "float"
                }
              }
            },
            "nullable": true
          },
          "default_prompt_suffix": {
            "type": "string",
            "nullable": true
          },
          "default_sd_params": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "エラーメッセージ"
          },
          "detail": {
            "type": "string",
            "description": "詳細情報"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "不正なリクエスト",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "リソースが見つかりません",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "内部サーバーエラー",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
...
}